Fixed cast that was giving this warning
cxalloc.cpp: In function `void* cvAlloc(size_t)':
cxalloc.cpp:105: warning: comparison is always false due to limited range of data type


git-svn-id: file://localhost/home/tolik/svnrepo/trunk@601 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ opencv/src/cxcore/cxalloc.cpp
@@ -102,7 +102,7 @@ CV_IMPL  void*  cvAlloc( size_t size )
 
     __BEGIN__;
 
-    if( (unsigned)size > CV_MAX_ALLOC_SIZE )
+    if( (size_t)size > CV_MAX_ALLOC_SIZE )
         CV_ERROR( CV_StsOutOfRange,
                   "Negative or too large argument of cvAlloc function" );
